pyproject.toml
[tool.poetry]
name = "asset inventory"
version = "0.1.0"
description = "For IT Asset Management/Tracking"
authors = ["Jaime Kingston"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.3"
sqlalchemy = "^2.0.25"
psycopg2 = "^2.9.9"
bcrypt = "^4.1.2"
black = "^23.12.1"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
